+++ src/script/figure.rs
@@ -20,27 +20,38 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 use std::sync::Arc;
 
+use serde::Serialize;
+
 use crate::generator::expression::{CircleExpr, Expression, LineExpr, PointExpr, ScalarExpr};
 
 use super::unroll::LabelMeta;
 
 type Point = Arc<Expression<PointExpr>>;
 
+/// The display mode of the expression.
+#[derive(Debug, Clone, Serialize, Copy)]
+pub enum Mode {
+    Dotted,
+    Dashed,
+    Bolded,
+    Default, // Normal solid curve
+}
+
 /// Defines the visual data of the figure.
 #[derive(Debug, Default)]
 pub struct Figure {
-    /// The points to be displayed
+    /// Points to be displayed
     pub points: Vec<(Arc<Expression<PointExpr>>, LabelMeta)>,
-    /// The lines to be displayed
-    pub lines: Vec<Arc<Expression<LineExpr>>>,
+    /// Lines to be displayed
+    pub lines: Vec<(Arc<Expression<LineExpr>>, Mode)>,
     /// Angles to be displayed
-    pub angles: Vec<(Arc<Expression<ScalarExpr>>, u8)>, // This u8 refers to number of arcs in an angle!
+    pub angles: Vec<(Arc<Expression<ScalarExpr>>, u8, Mode)>, // This u8 refers to number of arcs in an angle!
     /// Segments to be displayed
-    pub segments: Vec<(Point, Point)>,
+    pub segments: Vec<(Point, Point, Mode)>,
     /// Rays to be displayed
-    pub rays: Vec<(Point, Point)>,
+    pub rays: Vec<(Point, Point, Mode)>,
     /// Circles to be displayed
-    pub circles: Vec<Arc<Expression<CircleExpr>>>,
+    pub circles: Vec<(Arc<Expression<CircleExpr>>, Mode)>,
     /// The canvas size.
     pub canvas_size: (usize, usize),
 }